+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_sessionpool.py
@@ -376,3 +376,113 @@ class ContainerappSessionPoolTests(ScenarioTest):
         # List Session Pools
         sessionpool_list = self.cmd("containerapp sessionpool list -g {}".format(resource_group)).get_output_in_json()
         self.assertTrue(len(sessionpool_list) == 0)
+
+    @AllowLargeResponse(8192)
+    @ResourceGroupPreparer()
+    @SubnetPreparer(location=TEST_LOCATION, delegations='Microsoft.App/environments',
+                    service_endpoints="Microsoft.Storage.Global")
+    def test_containerapp_sessionpool_oncontainerexit(self, resource_group, subnet_id, vnet_name, subnet_name):
+        location = TEST_LOCATION
+        self.cmd('configure --defaults location={}'.format(location))
+
+        env_name = self.create_random_name(prefix='aca-sp-env', length=24)
+        self.cmd('containerapp env create -g {} -n {} -l {} --logs-destination none -s {}'.format(resource_group, env_name, location, subnet_id), expect_failure=False)
+        containerapp_env = self.cmd('containerapp env show -g {} -n {}'.format(resource_group, env_name)).get_output_in_json()
+        while containerapp_env["properties"]["provisioningState"].lower() in ["waiting", "inprogress"]:
+            time.sleep(5)
+            containerapp_env = self.cmd('containerapp env show -g {} -n {}'.format(resource_group, env_name)).get_output_in_json()
+
+        # List Session Pools
+        sessionpool_list = self.cmd("containerapp sessionpool list -g {}".format(resource_group)).get_output_in_json()
+        self.assertTrue(len(sessionpool_list) == 0)
+
+        # Create CustomContainer SessionPool
+        sessionpool_name_custom = self.create_random_name(prefix='spcustomcontainer', length=24)
+        ready_instances = 2
+        image = "mcr.microsoft.com/k8se/quickstart:latest"
+        secret_name = "testsecret"
+        secret_value = "testsecretvalue"
+        cpu = "0.5"
+        memory = "1Gi"
+        lifecycle_type = "OnContainerExit"
+        max_alive_period = 600
+        self.cmd(
+            'containerapp sessionpool create -g {} -n {} --container-type CustomContainer --environment {} --secrets {}={} --ready-sessions {} --image {} --cpu {} --memory {} --target-port {} --lifecycle-type {} --max-alive-period {}'.format(
+                resource_group, sessionpool_name_custom, env_name, secret_name, secret_value, ready_instances, image, cpu, memory, 80, lifecycle_type, max_alive_period),
+            checks=[
+                JMESPathCheck('name', sessionpool_name_custom),
+                JMESPathCheck('properties.containerType', "CustomContainer"),
+                JMESPathCheck('properties.scaleConfiguration.maxConcurrentSessions', 10),
+                JMESPathCheck('properties.scaleConfiguration.readySessionInstances', ready_instances),
+                JMESPathCheck('properties.secrets[0].name', secret_name),
+                JMESPathCheck('properties.customContainerTemplate.containers[0].image', image),
+                JMESPathCheck('properties.customContainerTemplate.containers[0].resources.cpu', cpu),
+                JMESPathCheck('properties.customContainerTemplate.containers[0].resources.memory', memory),
+                JMESPathCheck('properties.customContainerTemplate.ingress.targetPort', 80),
+                JMESPathCheck('properties.dynamicPoolConfiguration.lifecycleConfiguration.lifecycleType', lifecycle_type),
+                JMESPathCheck('properties.dynamicPoolConfiguration.lifecycleConfiguration.maxAlivePeriodInSeconds', max_alive_period),
+            ])
+
+        sessionpool = self.cmd('containerapp sessionpool show -g {} -n {}'.format(resource_group, sessionpool_name_custom)).get_output_in_json()
+        while sessionpool["properties"]["provisioningState"].lower() in ["waiting", "inprogress"]:
+            time.sleep(5)
+            sessionpool = self.cmd('containerapp sessionpool show -g {} -n {}'.format(resource_group, sessionpool_name_custom)).get_output_in_json()
+
+        self.cmd(
+            'containerapp sessionpool show -g {} -n {}'.format(resource_group, sessionpool_name_custom),
+            checks=[JMESPathCheck('properties.provisioningState', "Succeeded")]
+        )
+
+        # List Session Pools
+        sessionpool_list = self.cmd("containerapp sessionpool list -g {}".format(resource_group)).get_output_in_json()
+        self.assertTrue(len(sessionpool_list) == 1)
+
+        # Update Session Pool
+        max_concurrent_session = 12
+        cpu = "2.0"
+        memory = "4Gi"
+        egress = "EgressDisabled"
+        max_alive_period = 400
+        secret_name = "testsecret1"
+        secret_value = "testsecretvalue1"
+        self.cmd(
+            'containerapp sessionpool update -g {} -n {} --max-sessions {} --cpu {} --memory {} --network-status {} --max-alive-period {} --secrets {}={}'.format(
+                resource_group, sessionpool_name_custom, max_concurrent_session, cpu, memory, egress, max_alive_period, secret_name, secret_value),
+            checks=[
+                JMESPathCheck('name', sessionpool_name_custom),
+                JMESPathCheck('properties.containerType', "CustomContainer"),
+                JMESPathCheck('properties.scaleConfiguration.maxConcurrentSessions', max_concurrent_session),
+                JMESPathCheck('properties.scaleConfiguration.readySessionInstances', ready_instances),
+                JMESPathCheck('properties.secrets[0].name', secret_name),
+                JMESPathCheck('properties.customContainerTemplate.containers[0].image', image),
+                JMESPathCheck('properties.customContainerTemplate.containers[0].resources.cpu', cpu),
+                JMESPathCheck('properties.customContainerTemplate.containers[0].resources.memory', memory),
+                JMESPathCheck('properties.customContainerTemplate.ingress.targetPort', 80),
+                JMESPathCheck('properties.sessionNetworkConfiguration.status', egress),
+                JMESPathCheck('properties.dynamicPoolConfiguration.lifecycleConfiguration.lifecycleType', lifecycle_type),
+                JMESPathCheck('properties.dynamicPoolConfiguration.lifecycleConfiguration.maxAlivePeriodInSeconds', max_alive_period),
+            ])
+
+        sessionpool = self.cmd('containerapp sessionpool show -g {} -n {}'.format(resource_group, sessionpool_name_custom)).get_output_in_json()
+        while sessionpool["properties"]["provisioningState"].lower() in ["waiting", "inprogress"]:
+            time.sleep(5)
+            sessionpool = self.cmd('containerapp sessionpool show -g {} -n {}'.format(resource_group, sessionpool_name_custom)).get_output_in_json()
+
+        self.cmd(
+            'containerapp sessionpool show -g {} -n {}'.format(resource_group, sessionpool_name_custom),
+            checks=[JMESPathCheck('properties.provisioningState', "Succeeded")]
+        )
+
+        # Show a Session Pool
+        self.cmd(
+            'containerapp sessionpool show -g {} -n {}'.format(resource_group, sessionpool_name_custom),
+            checks=[
+                JMESPathCheck('name', sessionpool_name_custom),
+                JMESPathCheck('properties.containerType', "CustomContainer")
+            ])
+
+        self.cmd('containerapp sessionpool delete -g {} -n {} --yes'.format(resource_group, sessionpool_name_custom))
+
+        # List Session Pools
+        sessionpool_list = self.cmd("containerapp sessionpool list -g {}".format(resource_group)).get_output_in_json()
+        self.assertTrue(len(sessionpool_list) == 0)